split extension functionality
