+++ .prettierrc
@@ -1,3 +1,3 @@
 {
-    "printWidth": 150
+    "printWidth": 200
 }
